requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.31.0
langchain>=0.2.0
langchain-community>=0.0.10
faiss-gpu>=1.7.4; platform_system != "Darwin"
faiss-cpu>=1.7.4; platform_system == "Darwin"
sentence-transformers>=2.2.2

# Web scraping
beautifulsoup4>=4.12.2
requests>=2.31.0

# Data processing
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2

# Web API
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
jinja2>=3.1.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1